clarify valid values for LSR_PUBLISH_COVERAGE
There are 4 valid values for LSR_PUBLISH_COVERAGE.  If the value
is empty or unset, the default value, no coverage will be published.
A value of `strict` means situations like missing data to be reported
are treated as errors.
A value of `debug` means to use `coveralls debug`
A value of `normal` means coverage results will be reported normally.
Any other value will abort the test run with an error.